+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Breaking Changes
     - Minimum Java version now 11
+    - `WorkerPool::halt` and `WorkerPool::drainAndHalt` will now permanently halt the `WorkerPool`
     - `WorkProcessor::halt` and `WorkProcessor::haltLater` will now permanently halt the `WorkProcessor`
 
 ## 3.4.2